+++ example/schema.py
@@ -1,3 +1,5 @@
-from simple_graphql.django import build_schema
+from simple_graphql.django import SchemaBuilder
+from simple_graphql.django.types import ModelClass
 
-schema = build_schema()
+schema_builder: SchemaBuilder = SchemaBuilder[ModelClass]()
+schema = schema_builder.build_schema()