+++ simple_graphql/django/__init__.py
@@ -1,11 +1,7 @@
-from .decorators import graphql_model, register_graphql_model
-from .schema import build_schema, schema_builder
+from .schema import SchemaBuilder
 from .types import ModelSchemaConfig
 
 __all__ = [
-    "graphql_model",
-    "register_graphql_model",
-    "build_schema",
-    "schema_builder",
+    "SchemaBuilder",
     "ModelSchemaConfig",
 ]