+++ simple_graphql/django/schema/utils.py
@@ -1,5 +1,5 @@
-from simple_graphql.django.types import ModelClass
+from typing import Type
 
 
-def get_node_name(model_cls: ModelClass) -> str:
+def get_node_name(model_cls: Type) -> str:
     return f"{model_cls.__name__}"